refactor(ci): factorise code for commenting on PR
