+++ .github/workflows/pr.yml
@@ -41,6 +41,9 @@ jobs:
         with:
           node-version: 18
       - run: npm ci
-      - run: npm run test
+      - run: npm run test-ci
         env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          BASE_SHA: ${{ github.event.pull_request.base.sha }}
+          GITHUB_SHA: ${{ github.sha }}
           OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
