+++ package.json
@@ -11,6 +11,7 @@
     "start": "ts-node ./src/index.ts review",
     "start-ci": "ts-node ./src/index.ts review --ci",
     "test": "ts-node ./src/index.ts test",
+    "test-ci": "ts-node ./src/index.ts test --ci",
     "build": "node utils/build.js",
     "postbuild": "node utils/shebang.js && chmod +x ./dist/index.js"
   },
