+++ src/common/model/AIModel.ts
