+++ src/test/constants.ts
@@ -13,3 +13,6 @@ Return the content of a valid typescript file that would pass the test case.
 
 // The threshold for the similarity score to pass the test.
 export const testThreshold = 0.1;
+
+export const signOff =
+  "#### Tests Powered by [Code Review GPT](https://github.com/mattzcarey/code-review-gpt)";
